+++ src/asyncore/mod.rs
@@ -1,0 +1,2 @@
+pub use async_std::prelude::*;
+pub use async_std::*;