+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v0.1.0
+
+- Added asyncore module
+- Added parser module
+- Initialized i18n module
+
 ### v0.0.0
 
 - init